requirements.txt
Flask==0.10.1
Flask-DebugToolbar==0.10.0
Flask-Login==0.2.11
Flask-Migrate==1.4.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-Session==0.2
#Flask-SuperAdmin==1.7.1
Flask-WTF==0.11
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
SQLAlchemy==1.0.3
#SQLAlchemy-FullText-Search==0.2.3
WTForms==2.0.2
Werkzeug==0.9.6
alembic==0.7.6
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.20
blinker==1.3
celery==3.1.18
itsdangerous==0.24
kombu==3.0.26
mysql-connector-python-rf==2.0.4
python-gcm==0.1.5
pytz==2015.4
redis==2.10.3
requests==2.7.0
wsgiref==0.1.2
xlrd==0.9.3
openpyxl==2.3.0-b1
